Make runner extensible
Create a new entry-point to register runner. A runner is just a function
that receive job definition and must exit the program.

The runner is defined from environment variable, controlled by the
sysadm. The default runner is `unconfined` which execute the test script
right in the CI env. The `unconfined` runner even allow to choose the
runner from the build.